use conftest for generic config fixture
